+++ lib/cjs/types/components/controls/Input/Input.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react" />
+import { InputProps } from "./Input.types";
+export declare const Input: ({ id, className, value, placeholder, type, width, size, view, disabled, onChange, }: InputProps) => JSX.Element;