+++ lib/cjs/types/infra/index.d.ts
@@ -1,1 +1,1 @@
-export { Theme } from "./Theme";
+export * from "./Theme";