+++ lib/cjs/types/shared/hooks/Theme.d.ts
@@ -5,38 +5,88 @@ export declare const useTheme: () => {
             button: {
                 primary: {
                     background: string;
-                    color: string;
+                    text: string;
                 };
                 secondary: {
                     background: string;
-                    color: string;
+                    text: string;
                 };
                 success: {
                     background: string;
-                    color: string;
+                    text: string;
                 };
                 danger: {
                     background: string;
-                    color: string;
+                    text: string;
                 };
                 warning: {
                     background: string;
-                    color: string;
+                    text: string;
                 };
                 info: {
                     background: string;
-                    color: string;
+                    text: string;
                 };
                 light: {
                     background: string;
-                    color: string;
+                    text: string;
                 };
                 dark: {
                     background: string;
-                    color: string;
+                    text: string;
+                };
+            };
+            input: {
+                primary: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
+                };
+                secondary: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
+                };
+                success: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
+                };
+                danger: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
+                };
+                warning: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
+                };
+                info: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
+                };
+                light: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
+                };
+                dark: {
+                    border: string;
+                    background: string;
+                    text: string;
+                    placeholder: string;
                 };
             };
         };
     };
-    setTheme: (name: string) => void;
+    setTheme: (theme: any) => void;
 };