+++ lib/esm/types/components/controls/Input/Input.styled.d.ts
@@ -1,0 +1,5 @@
+import { InputProps } from "./Input.types";
+export declare const StyledInput: import("styled-components").StyledComponent<"div", any, InputProps, never>;
+export declare const StyledInputContent: import("styled-components").StyledComponent<"input", any, {
+    fontSize?: InputProps["size"];
+} & InputProps, never>;