+++ lib/esm/types/infra/Theme/Theme.d.ts
@@ -1,3 +1,3 @@
 /// <reference types="react" />
 import { ThemeProps } from "./Theme.types";
-export declare const Theme: ({ children }: ThemeProps) => JSX.Element;
+export declare const Theme: ({ children, theme }: ThemeProps) => JSX.Element;