+++ lib/esm/types/infra/Theme/Theme.types.d.ts
@@ -1,4 +1,9 @@
 import { ReactNode } from "react";
 export declare type ThemeProps = {
+    theme?: any;
     children: ReactNode;
+};
+export declare type ThemeContextProviderProps = {
+    theme?: any;
+    setTheme?: (theme: any) => void;
 };