+++ src/infra/Theme/Theme.tsx
@@ -1,15 +1,13 @@
 import React from "react";
 import { ThemeProvider } from "styled-components";
-import { useTheme, GlobalStyle } from "../..";
+import { defaultTheme, GlobalStyle } from "../..";
 import { ThemeProps } from "./Theme.types";
 
-export const Theme = ({ children }: ThemeProps) => {
-  const { theme } = useTheme();
-
+export const Theme = ({ children, theme }: ThemeProps) => {
   return (
     <>
       <GlobalStyle />
-      <ThemeProvider theme={theme}>{children}</ThemeProvider>
+      <ThemeProvider theme={theme || defaultTheme}>{children}</ThemeProvider>
     </>
   );
 };