+++ src/shared/hooks/Theme.tsx
@@ -1,16 +1,11 @@
-import { useState } from "react";
-import { defaultTheme, themes } from "../styles";
+import React, { useState } from "react";
+import { defaultTheme } from "../styles";
 
 export const useTheme = () => {
   const [currentTheme, setCurrentTheme] = useState(defaultTheme);
 
-  const setTheme = (name: string) => {
-    const theme = themes.find((item) => item.name === name);
-    if (theme?.name === name) {
-      setCurrentTheme(theme);
-    } else {
-      setCurrentTheme(defaultTheme);
-    }
+  const setTheme = (theme: any) => {
+    setCurrentTheme(theme);
   };
 
   return { theme: currentTheme, setTheme };