+++ src/shared/styles/globalStyles.ts
@@ -8,5 +8,6 @@ export const GlobalStyle = createGlobalStyle`
   }
   * {
     font-family: 'Roboto', sans-serif;
+    outline: none;
   }
 `;