+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Changelog
 
+## v1.4.0 (02/05/2022)
+Important thinks:
+
+- The Input component was created&.
+--------------------
+
+## v1.3.0 (02/05/2022)
+Important thinks:
+
+- The bug with the `useTheme` was fixed.
+- The `Button` component was updated.
+--------------------
+
 ## v1.2.1 (02/03/2022)
 Important thinks:
 
-- The bug with `defaultTheme` was fixed'
-- The `Button` was updated
-- The hook `useTheme` was created 
+- The bug with `defaultTheme` was fixed'.
+- The `Input` was updated.
+- The hook `useTheme` was created.
 --------------------
 
 ## v1.2.0 (02/03/2022)
@@ -17,5 +30,5 @@ Important thinks:
 ## v1.1.0 (02/02/2022)
 Important thinks:
 
-- The `Button` component was updated.
+- The `Input` component was updated.
 --------------------